requirements.txt
python-telegram-bot==20.7
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.0
aiohttp==3.9.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
requests==2.31.0